// src/app/main/apps/procesos/contratacion/models/contratacion.model.ts
export class EtapaResponseModel {
    idEtapa: number;
    idProceso: number;
    anio: number;
    idEtapaFase?: number;
    codigoProceso: string;
    descripcionEtapa: string;
    descripcionRegimenLaboral: string;
    descripcionEtapaFase: string;
    descripcionNumeroConvocatoria: string;
    descripcionTipoProceso: string;
    fechaCreacion: string;
    descripcionNivelEducativo?: string;
    descripcionCargosConvocados?: string;
    idRegimenLaboral?: number;
    idNivelEducativo?: number;
    codigoRegimenLaboral?: number;
    codigoEtapaFase: number;
    codigoEtapa: string;
    descripcioncargosconvocados?: string;
    descripcionEstadoEtapaProceso: string;
    idEstadoDesarrollo?: number;
    // adicionales
    accesoPublicacionPlaza?: boolean;
    accesoValidarPlazas?: boolean;
    accesoAprobarPlazas?: boolean;
    accesoRechazarPlazas?: boolean;
    accesoPublicarPlazas?: boolean;
    accesoPostulantes?: boolean;
    accesoCargaMasivaPostulantes?: boolean;
    accesoEliminarCargaPostulantes?: boolean;
    accesoCalificacion?: boolean;
    accesoPublicarCalificacion?: boolean;
    accesoCargaMasivaCalificacion?: boolean;
    accesoEliminarCargaCalificacion?: boolean;
    accesoCalificacionAutomatica?: boolean;
    accesoAdjudicacion?: boolean;
    accesoPublicarAdjudicados?: boolean;
    accesoCargaMasivaAdjudicacion?: boolean;
    accesoEliminarCargaAdjudicacion?: boolean;
    accesoConsolidado?: boolean;
    accesoAprobarPlazaConsolidado?: boolean;
    accesoPlazaPrepublicada?: boolean;
    accesoPlazaResultadoFinal?: boolean;
    accesoRegistrarCalificacion?: boolean;
    accesoRegistrarReclamoCalificacion?: boolean;
    accesoFinalizarEtapa?: boolean;
    accesoAdjudicarPlaza?: boolean;
    accesoNoAdjudicarPlaza?: boolean;
}
// ...
export class ComboItem {
    constructor(value: number, label: string) {
        this.value = value;
        this.label = label;
    }
    value: number;
    label: string;
}

export class OpcionFiltro {
    item = new ComboItem(-1, "TODOS");
}